FlaskApp/requirements.txt
beautifulsoup4==4.9.3
click==7.1.2
Flask==1.1.2
google==3.0.0
grpcio==1.37.1
grpcio-tools==1.37.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
nltk==3.6.2
protobuf==3.15.8
regex==2021.4.4
six==1.15.0
soupsieve==2.2.1
tqdm==4.60.0
Werkzeug==1.0.1
